Split api spec into separate class
