+++ peek/completer.py
@@ -1,9 +1,7 @@
-import ast
 import itertools
-import json
 import logging
 import os
-from typing import Iterable, List, Dict, Tuple, Optional
+from typing import Iterable, List, Tuple, Optional
 
 from prompt_toolkit.completion import Completer, CompleteEvent, Completion, WordCompleter, FuzzyCompleter, PathCompleter
 from prompt_toolkit.contrib.completers import SystemCompleter
@@ -12,10 +10,9 @@ from pygments.token import Error, Name, Literal, String
 
 from peek.common import PeekToken
 from peek.completions import PayloadKeyCompletion
-from peek.errors import PeekError
-from peek.es_api_spec.spec_builder import build_js_specs
+from peek.es_api_spec.spec import ApiSpec
 from peek.lexers import PeekLexer, UrlPathLexer, PathPart, ParamName, Ampersand, QuestionMark, Slash, HttpMethod, \
-    FuncName, OptionName, Assign, CurlyLeft, CurlyRight, DictKey, ShellOut, At
+    FuncName, OptionName, Assign, DictKey, ShellOut, At
 from peek.parser import process_tokens
 
 _logger = logging.getLogger(__name__)
@@ -37,17 +34,7 @@ class PeekCompleter(Completer):
         from peek import __file__ as package_root
         package_root = os.path.dirname(package_root)
         kibana_dir = app.config['kibana_dir'] or os.path.join(package_root, 'specs', 'kibana-7.8.1')
-        if not self.app.batch_mode and self.app.config.as_bool('load_api_specs'):
-            _logger.info(f'Build API specs from: {kibana_dir}')
-            self.specs = load_specs(kibana_dir)
-            if self.app.config.as_bool('load_extended_api_specs'):
-                _logger.info(f'Build extended API specs from: {kibana_dir}')
-                self.specs = _merge_specs(self.specs, build_js_specs(kibana_dir))
-        else:
-            self.specs = {}
-
-        with open('tmp-all-specs.json', 'w') as outs:
-            json.dump(self.specs, outs, indent=2)
+        self.api_spec = ApiSpec(app, kibana_dir)
 
     def get_completions(self, document: Document, complete_event: CompleteEvent) -> Iterable[Completion]:
         _logger.debug(f'doc: {document}, event: {complete_event}')
@@ -138,7 +125,7 @@ class PeekCompleter(Completer):
         method = method_token.value.upper()
         cursor_position = document.cursor_position - path_token.index
         path = path_token.value[:cursor_position]
-        _logger.debug(f'Completing http path: {path!r}')
+        _logger.debug(f'Completing HTTP API url: {path!r}')
         path_tokens = list(self.url_path_lexer.get_tokens_unprocessed(path))
         _logger.debug(f'path_tokens: {path_tokens}')
         if not path_tokens:  # empty, should not happen
@@ -148,65 +135,15 @@ class PeekCompleter(Completer):
         _logger.debug(f'cursor_token: {cursor_token}')
         if cursor_token.ttype is Error:
             return []
-        elif cursor_token.ttype in (PathPart, Slash):
-            return self._complete_http_path_part(document, complete_event, method, path_tokens)
-        elif cursor_token.ttype in (ParamName, QuestionMark, Ampersand):
-            return self._complete_query_param_name(document, complete_event, method, path_tokens)
-        else:  # skip for param value
-            return self._complete_query_param_value(document, complete_event, method, path_tokens)
-
-    def _complete_http_path_part(self, document: Document, complete_event: CompleteEvent, method, path_tokens):
-        cursor_token = path_tokens[-1]
-        _logger.debug(f'Completing path part: {cursor_token}')
-        ts = [t.value for t in path_tokens if t.ttype is not Slash]
-        if cursor_token.ttype is PathPart:
-            ts.pop()
-        _logger.debug(f'ts: {ts}')
-        candidates = []
-        for api_name, api_spec in self.specs.items():
-            if 'methods' not in api_spec:
-                continue
-            if method not in api_spec['methods']:
-                continue
-            for api_path in api_spec['patterns']:
-                ps = [p for p in api_path.split('/') if p]
-                # Nothing to complete if the candidate is shorter than current input
-                if len(ts) >= len(ps):
-                    continue
-                if not can_match(ts, ps):
-                    continue
-                candidate = '/'.join(ps[len(ts):])
-                candidates.append(Completion(candidate, start_position=0))
-
-        return FuzzyCompleter(ConstantCompleter(candidates)).get_completions(document, complete_event)
-
-    def _complete_query_param_name(self, document: Document, complete_event: CompleteEvent, method, path_tokens):
-        _logger.debug(f'Completing query param name: {path_tokens[-1]}')
-        ts = [t.value for t in path_tokens if t.ttype is PathPart]
-        candidates = set()
-        for api_spec in matchable_specs(method, ts, self.specs):
-            candidates.update(api_spec['url_params'].keys())
-
-        return FuzzyCompleter(ConstantCompleter(
-            [Completion(c, start_position=0) for c in candidates])).get_completions(document, complete_event)
-
-    def _complete_query_param_value(self, document: Document, complete_event: CompleteEvent, method, path_tokens):
-        _logger.debug(f'Completing query param value: {path_tokens[-1]}')
-        param_name_token = path_tokens[-2] if path_tokens[-1].ttype is Assign else path_tokens[-3]
-        _logger.debug(f'Param name token: {param_name_token}')
-        ts = [t.value for t in path_tokens if t.ttype is PathPart]
-        candidates = set()
-        for api_spec in matchable_specs(method, ts, self.specs):
-            v = api_spec['url_params'].get(param_name_token.value, None)
-            if v is None:
-                continue
-            if v == '__flag__':
-                candidates.update(('true', 'false'))
-            elif isinstance(v, list) and len(v) > 0:
-                candidates.update(v)
-
-        return FuzzyCompleter(ConstantCompleter(
-            [Completion(c, start_position=0) for c in candidates])).get_completions(document, complete_event)
+        else:
+            if cursor_token.ttype in (PathPart, Slash):
+                complete_func = self.api_spec.complete_url_path
+            elif cursor_token.ttype in (ParamName, QuestionMark, Ampersand):
+                complete_func = self.api_spec.complete_query_param_name
+            else:
+                complete_func = self.api_spec.complete_query_param_value
+            candidates = complete_func(document, complete_event, method, path_tokens)
+            return FuzzyCompleter(ConstantCompleter(candidates)).get_completions(document, complete_event)
 
     def _complete_options(self, document: Document, complete_event: CompleteEvent, tokens: List[PeekToken],
                           is_cursor_on_non_white_token):
@@ -247,128 +184,16 @@ class PeekCompleter(Completer):
         _logger.debug(f'Completing for payload with tokens: {tokens}')
         method_token, path_token = tokens[0], tokens[1]
         path_tokens = list(self.url_path_lexer.get_tokens_unprocessed(path_token.value))
-        ts = [t.value for t in path_tokens if t.ttype is not Slash]
-
-        try:
-            api_spec = next(matchable_specs(method_token.value.upper(), ts, self.specs,
-                                            required_field='data_autocomplete_rules'))
-        except StopIteration:
-            return []
-
-        rules = self._maybe_process_rules(api_spec.get('data_autocomplete_rules', None))
-        if rules is None:
-            return []
-
-        _logger.debug(f'Found rules from spec: {rules}')
-
-        payload_keys = []
-        curly_level = 0
-        for t in tokens[2:-1]:
-            if t.ttype is CurlyLeft:
-                curly_level += 1
-            elif t.ttype is CurlyRight:
-                curly_level -= 1
-                payload_keys.pop()
-            elif t.ttype is DictKey:
-                if len(payload_keys) == curly_level - 1:
-                    payload_keys.append(ast.literal_eval(t.value))
-                elif len(payload_keys) == curly_level:
-                    payload_keys[-1] = ast.literal_eval(t.value)
-                else:
-                    raise PeekError(f'Error when counting curly level {curly_level} and keys {payload_keys}')
-
-        _logger.debug(f'Payload status: level: {curly_level}, keys: {payload_keys}')
-        if curly_level == 0:  # not even in the first curly bracket, no completion
-            return []
-
-        # Remove the payload key that is at the same level
-        if curly_level == len(payload_keys):
-            payload_keys.pop()
-
-        rules = self._resolve_rules_for_keys(rules, payload_keys)
-        if rules is None:
-            _logger.debug(f'Rules not available for key: {payload_keys!r}')
-            return []
-
-        # TODO: handle __scope_link
-        # TODO: __one_of, e.g. POST _render/template
-        # TODO: '*' matching
-        # TODO: top-level __template, e.g. POST _reindex
-        # TODO: filters how does it work
-        constant_completer = ConstantCompleter([Completion(k, start_position=0) for k in rules.keys()
-                                                if k not in ('__scope_link', '__template', '__one_of')])
+        # TODO: this is not correct since there maybe options after the first 2 tokens
+        #       In fact, there is no way to tell for sure where the payload begins without
+        #       parse the token stream
+        payload_tokens = tokens[2:]
+        candidates, rules = self.api_spec.complete_payload(
+            document, complete_event, method_token.value.upper(), path_tokens, payload_tokens)
+        constant_completer = ConstantCompleter(candidates)
         for c in FuzzyCompleter(constant_completer).get_completions(document, complete_event):
             yield PayloadKeyCompletion(c.text, rules[c.text],
                                        c.start_position, c.display, c.display_meta, c.style, c.selected_style)
-
-    def _resolve_rules_for_keys(self, rules, payload_keys):
-        for i, k in enumerate(payload_keys):
-            rules = self._maybe_process_rules(rules.get(k, None))
-            # Special handle for query
-            if k == 'query' and rules == {}:
-                rules = self.specs['GLOBAL']['query']
-            if rules is None:
-                break
-        return rules
-
-    def _maybe_process_rules(self, rules):
-        processors = [
-            self._maybe_resolve_scope_link,
-            self._maybe_unwrap_for_dict,
-            self._maybe_lift_one_of,
-        ]
-        while True:
-            original_rules = rules
-            for p in processors:
-                rules = p(rules)
-            if rules == original_rules:
-                break
-        return rules
-
-    def _maybe_resolve_scope_link(self, rules):
-        if isinstance(rules, dict) and '__scope_link' in rules:
-            rules = dict(rules)  # avoid mutating original value
-            _logger.debug(f'Found scope link: {rules!r}')
-            scope_link = rules.pop('__scope_link')
-            if scope_link.startswith('.'):  # TODO: relative scope link
-                _logger.debug('Relative scope link not implemented')
-            elif '.' in scope_link:
-                scope = self.specs
-                for scope_link_key in scope_link.split('.'):
-                    if scope_link_key in scope:
-                        scope = scope[scope_link_key]
-                if id(scope) != id(self.specs):
-                    rules.update(scope['data_autocomplete_rules'] if 'data_autocomplete_rules' in scope else scope)
-            else:
-                scope = self.specs.get(scope_link, None)
-                if isinstance(scope, dict):
-                    rules.update(scope['data_autocomplete_rules'] if 'data_autocomplete_rules' in scope else scope)
-        return rules
-
-    def _maybe_lift_one_of(self, rules):
-        if isinstance(rules, dict) and '__one_of' in rules:
-            rules = dict(rules)
-            one_of = rules.pop('__one_of')
-            if isinstance(one_of[0], dict):
-                for e in one_of:
-                    rules.update(e)
-            else:
-                rules['__one_of'] = one_of
-        return rules
-
-    def _maybe_unwrap_for_dict(self, rules):
-        """
-        If the rules is an list of dict, return the first dict
-        """
-        if isinstance(rules, dict):
-            return rules
-        elif isinstance(rules, list) and len(rules) > 0:
-            if isinstance(rules[0], dict):
-                return rules[0]
-            else:
-                return None
-        else:
-            return None
 
 
 def find_beginning_token(tokens) -> Tuple[Optional[int], Optional[PeekToken]]:
@@ -378,45 +203,6 @@ def find_beginning_token(tokens) -> Tuple[Optional[int], Optional[PeekToken]]:
     return None, None
 
 
-def matchable_specs(method: str, ts: List[str], specs: Dict, required_field='url_params') -> Dict:
-    """
-    Find the matchable spec for the given HTTP method and input path.
-    """
-    for api_name, api_spec in specs.items():
-        if 'methods' not in api_spec:
-            continue
-        if method not in api_spec['methods']:
-            continue
-        if not api_spec.get(required_field, None):
-            continue
-        matched = False
-        for pattern in api_spec['patterns']:
-            ps = [p for p in pattern.split('/') if p]
-            if len(ts) != len(ps):
-                continue
-            if not can_match(ts, ps):
-                continue
-            matched = True
-            break
-        if matched:
-            yield api_spec
-
-
-def can_match(ts, ps):
-    """
-    Test whether the input path (ts) can match the candidate path (ps).
-    The rule is basically a placeholder can match any string other than
-    the ones leading with underscore.
-    """
-    for t, p in zip(ts, ps):
-        if t != p:
-            if t.startswith('_'):
-                return False
-            if not (p.startswith('{') and p.endswith('}')):
-                return False
-    return True
-
-
 class ConstantCompleter(Completer):
 
     def __init__(self, candidates):
@@ -424,57 +210,3 @@ class ConstantCompleter(Completer):
 
     def get_completions(self, document: Document, complete_event: CompleteEvent) -> Iterable[Completion]:
         return self.candidates
-
-
-def load_specs(kibana_dir):
-    oss_path = os.path.join(
-        kibana_dir, 'src', 'plugins', 'console', 'server', 'lib', 'spec_definitions')
-    xpack_path = os.path.join(
-        kibana_dir, 'x-pack', 'plugins', 'console_extensions', 'server', 'lib', 'spec_definitions')
-    specs = _load_json_specs(os.path.join(oss_path, 'json'))
-    specs.update(_load_json_specs(os.path.join(xpack_path, 'json')))
-    return specs
-
-
-def _load_json_specs(base_dir):
-    _logger.debug(f'Loading json specs from: {base_dir!r}')
-    specs = {}
-    for sub_dir in ('generated', 'overrides'):
-        d = os.path.join(base_dir, sub_dir)
-        if not os.path.exists(d):
-            _logger.warning(f'JSON specs directory does not exist: {d}')
-            continue
-        for f in os.listdir(d):
-            if f == '_common.json':
-                continue
-            with open(os.path.join(d, f)) as ins:
-                spec = json.load(ins)
-            if sub_dir == 'generated':
-                specs.update(spec)
-            else:
-                for k, v in spec.items():
-                    if k in specs:
-                        specs[k].update(v)
-                    else:
-                        if k.startswith('xpack.'):
-                            specs[k[6:]].update(v)
-                        else:
-                            specs['xpack.' + k].update(v)
-
-    return specs
-
-
-def _merge_specs(basic_specs, extended_specs):
-    specs = dict(basic_specs)
-    for k, v in extended_specs.items():
-        if k not in specs:
-            specs[k] = v
-        else:
-            for kk, vv in v.items():
-                if kk in specs[k]:
-                    _logger.debug(f'Duplicated key found: {k}.{kk}')
-                    if isinstance(specs[k][kk], dict) and isinstance(vv, dict):
-                        specs[k][kk].update(vv)
-                else:
-                    specs[k][kk] = vv
-    return specs
